+++ hvega/CHANGELOG.md
@@ -1,6 +1,27 @@
 For the latest version of this document, please see
 [https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md](https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md).
 
+## 0.11.0.0
+
+The Vega-Lite tests are now validated against version 4.14 of the
+Vega-Lite schema.
+
+Note that `hvega` does __not__ provide any information to help users
+take advantage of the (new to 4.14) ability to
+omit the type of a field when
+[it can be inferred](https://vega.github.io/vega-lite/docs/type.html).
+As the type is currently optional in `hvega` users can just
+not give a type.
+
+
+### New Constructors
+
+The `OrderChannel` type has gained `OBand`, `OTitle`/`ONoTitle`,
+and conditional predicate support with `ODataCondition`,
+`OSelectionCondition`, and `ONumber` constructors.
+
+The `MarkChannel` type has gained the `MNullValue` constructor.
+
 ## 0.10.0.0
 
 The Vega-Lite tests are now validated against version 4.13 of the